[fix] as per convo with valerie, we can't have templater be separated from render because there could be two separate instances of handlebars
